refactor: modify AppState with ReturnType
