+++ src/components/pages/SettingPage/index.tsx
@@ -9,7 +9,7 @@ import Toggle from 'react-toggle';
 import 'react-toggle/style.css';
 import styled from 'styled-components';
 import { CountType } from '../../../models/consts/count-type';
-import { AppState } from '../../../redux/slices/app-state';
+import { AppState } from '../../../redux/app-reducer';
 import { DeleteBlogState, deleteBlogSlice, deleteBlog } from '../../../redux/slices/delete-blog-state';
 import { FeedState } from '../../../redux/slices/feeds-state';
 import Anker from '../../atoms/Anker/index';
