+++ src/components/pages/SignInPage/index.tsx
@@ -8,7 +8,7 @@ import { StyledFirebaseAuth } from 'react-firebaseui';
 import { useDispatch, useSelector } from 'react-redux';
 import { Redirect, RouteComponentProps } from 'react-router-dom';
 import { UserState, fetchUser } from '../../../redux/slices/user-state';
-import { AppState } from '../../../redux/slices/app-state';
+import { AppState } from '../../../redux/app-reducer';
 import Anker from '../../atoms/Anker/index';
 import Wrapper from '../../atoms/Wrapper/index';
 import LoadingView from '../../molecules/LoadingView/index';
